# workspace/cpg.bin/workspace/cpg.bin/workspace/cpg.bin/workspace/cpg.bin/workspace/cpg.bin/workspace/cpg.bin/workspace/cpg.bin/workspace/cpg.bin/workspace/cpg.bin/server/api.yaml
openapi: "3.0.3"
info:
  title: Call Center Broker API
  version: "1.0.0"
  description: >
    Broker API that dispatches owner leads to agent mobile devices.
    Agents authenticate, receive owners to call, and submit call outcomes.
servers:
  - url: https://api.server.test/api/v1
  - url: http://localhost:3000/api/v1
    description: Local dev
security:
  - BearerAuth: []
  # - apiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    ClientType:
      type: string
      enum: [OWNER, LEAD, BOTH]
    OwnerPhone:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
          example: "+201001234567"
    OwnerInfoEntry:
      type: object
      required: [key, value]
      properties:
        key:
          type: string
          example: "national_id"
        value:
          type: string
          example: "29901011234567"
    Owner:
      type: object
      required: [id]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: "Mohamed Ali"
        type:
          $ref: "#/components/schemas/ClientType"
        next_dial_at:
          type: string
          format: date-time
          nullable: true
        agent_id:
          type: integer
          nullable: true
        phones:
          type: array
          items:
            $ref: "#/components/schemas/OwnerPhone"
        info:
          type: array
          items:
            $ref: "#/components/schemas/OwnerInfoEntry"
    CreateOwnerRequest:
      type: object
      required: [phones]
      properties:
        name:
          type: string
        project_id:
          type: integer
        agent_id:
          type: integer
        type:
          $ref: "#/components/schemas/ClientType"
        phones:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OwnerPhone"
        info:
          type: array
          items:
            $ref: "#/components/schemas/OwnerInfoEntry"
    BulkCreateOwnersRequest:
      type: object
      required: [owners]
      properties:
        owners:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CreateOwnerRequest"
...
    BulkCreateUsersRequest:
      type: object
      required: [users]
      properties:
        users:
          type: array
          minItems: 1
          items:
            type: object
            required: [email, password, role]
            properties:
              email:
                type: string
              password:
                type: string
                format: password
              name:
                type: string
              phone:
                type: string
              role:
                type: string
    User:
      type: object
      required: [id, email, role]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        role:
          type: string
        has_profile_image:
          type: boolean
          description: Whether the user has a profile image stored
        is_online:
          type: boolean
...
    PaginatedOwners:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Owner"
        meta:
          type: object
          required: [total, page, limit]
          properties:
            total:
              type: integer
            page:
              type: integer
            limit:
              type: integer
    StatusCount:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
          example: "completed"
        count:
          type: integer
          example: 42
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "OWNER_NOT_FOUND"
        message:
          type: string
          example: "No owner with that ID exists"
        details:
          type: object
          nullable: true
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Request body or params failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
paths:
...
  /owners:
    get:
      summary: List owners (paginated, filterable)
      operationId: listOwners
      tags: [owners]
      parameters:
        - name: project_id
          in: query
          schema:
            type: integer
        - name: agent_id
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/ClientType"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Paginated owner list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOwners"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      summary: Create a new owner
      operationId: createOwner
      tags: [owners]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOwnerRequest"
      responses:
        "201":
          description: Owner created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
        "422":
          $ref: "#/components/responses/ValidationError"
  /owners/bulk:
    post:
      summary: Bulk create or merge owners
      operationId: bulkCreateOwners
      tags: [owners]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCreateOwnersRequest"
      responses:
        "201":
          description: Owners created or updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Owner"
        "422":
          $ref: "#/components/responses/ValidationError"
  /owners/{ownerId}:
    get:
      summary: Get a single owner with numbers and info
      operationId: getOwner
      tags: [owners]
      parameters:
        - name: ownerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Owner detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      summary: Update owner type or snooze (next_dial_at)
      operationId: updateOwner
      tags: [owners]
      parameters:
        - name: ownerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  $ref: "#/components/schemas/ClientType"
                agent_id:
                  type: integer
                  nullable: true
                next_dial_at:
                  type: string
                  format: date-time
                  nullable: true
      responses:
        "200":
          description: Updated owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete an owner
      operationId: deleteOwner
      tags: [owners]
      parameters:
        - name: ownerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Owner deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /owners/statuses:
    get:
      summary: Get distinct client statuses with counts (from client_project)
      operationId: getOwnerStatuses
      tags: [owners]
      responses:
        "200":
          description: List of statuses with counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatusCount"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /owners/{ownerId}/projects:
    post:
      summary: Assign an owner to a project by project name
      operationId: assignOwnerProject
      tags: [owners]
      parameters:
        - name: ownerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_name]
              properties:
                project_name:
                  type: string
      responses:
        "200":
          description: Owner updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /users:
    get:
      summary: List all users (admin only)
      operationId: listUsers
      tags: [users]
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: online
          in: query
          description: Filter by online status ('true' or 'false')
          schema:
            type: string
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      summary: Create a new user (admin only)
      operationId: createUser
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, role]
              properties:
                email:
                  type: string
                password:
                  type: string
                  format: password
                name:
                  type: string
                phone:
                  type: string
                role:
                  type: string
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
  /users/bulk:
    post:
      summary: Bulk create users (admin only)
      operationId: bulkCreateUsers
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCreateUsersRequest"
      responses:
        "201":
          description: Users created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Duplicate emails found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
  /users/{userId}:
    get:
      summary: Get a user profile
      operationId: getUser
      tags: [users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      summary: Update a user profile (self or admin)
      operationId: updateUser
      tags: [users]
      description: |
        Admins may update any user. A non-admin user may only update their own
        profile (name, phone, email, password). The `role` field is admin-only
        and is ignored for non-admin requests.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                  format: password
                name:
                  type: string
                phone:
                  type: string
                role:
                  type: string
                  description: Admin only. Ignored for non-admin self-updates.
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /users/{userId}/profile-image:
    get:
      summary: Get the profile image for a user
      operationId: getProfileImage
      tags: [users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile image binary
          content:
            image/*:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      summary: Upload a profile image for a user
      operationId: uploadProfileImage
      tags: [users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [profile_image]
              properties:
                profile_image:
                  type: string
                  format: binary
      responses:
        "200":
          description: Profile image uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete the profile image for a user
      operationId: deleteProfileImage
      tags: [users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile image deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{userId}/deactivate:
    patch:
      summary: Deactivate a user (admin only) — sets role to 'deactivated' and invalidates password
      operationId: deactivateUser
      tags: [users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{userId}/stats:
    get:
      summary: Get call statistics for an agent
      operationId: getUserStats
      tags: [users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Aggregated stats for this agent
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_calls:
                    type: integer
                  answered:
                    type: integer
                  no_answer:
                    type: integer
                  busy:
                    type: integer
                  failed:
                    type: integer
                  callback:
                    type: integer
                  avg_duration_seconds:
                    type: number
                    format: float
                  total_session_time_seconds:
                    type: integer
        "404":
          $ref: "#/components/responses/NotFound"